refactor(site): share machine doc category rows
