+++ site/src/lib/machine-docs.test.ts
@@ -1,7 +1,9 @@
 import assert from "node:assert/strict";
 import test from "node:test";
+import { totalCategories } from "../data/full-catalog.ts";
 import { renderIndexMarkdown } from "../pages/index.md.ts";
 import { renderLlmsText } from "../pages/llms.txt.ts";
+import { renderCategoryRows } from "./machine-docs.ts";
 
 const PREFIX = "https://example.test/";
 const CATEGORY_HEADER_PATTERN = /^categories\[14\]\{category,total,copa,wolfi\}:$/m;
@@ -37,4 +39,18 @@ test("renderIndexMarkdown exposes aggregates and contextual next actions", () =>
   assert.match(output, CATEGORY_HEADER_PATTERN);
   assert.ok(output.includes("## Next Actions"));
   assert.ok(!output.includes("Agent Interface"));
+});
+
+test("machine doc routes share the catalog category rows", () => {
+  // Given
+  const categoryRows = renderCategoryRows();
+
+  // When
+  const llmsText = renderLlmsText(PREFIX);
+  const indexMarkdown = renderIndexMarkdown(PREFIX);
+
+  // Then
+  assert.equal(categoryRows.split("\n").length, totalCategories);
+  assert.ok(llmsText.includes(categoryRows));
+  assert.ok(indexMarkdown.includes(categoryRows));
 });
