+++ site/src/pages/index.md.ts
@@ -1,22 +1,11 @@
 import type { APIRoute } from "astro";
-import {
-  copaCount,
-  fullCatalog,
-  integerCount,
-  totalCategories,
-  totalImages,
-} from "../data/full-catalog.ts";
+import { copaCount, integerCount, totalCategories, totalImages } from "../data/full-catalog.ts";
 import { getChartsCatalog } from "../lib/charts.ts";
+import { renderCategoryRows } from "../lib/machine-docs.ts";
 
 export function renderIndexMarkdown(prefix: string): string {
   const chartCount = getChartsCatalog().charts.length;
-  const categoryRows = fullCatalog
-    .map((category) => {
-      const copaImages = category.images.filter((image) => image.source === "copa").length;
-      const wolfiImages = category.images.length - copaImages;
-      return `  ${category.id},${category.images.length},${copaImages},${wolfiImages}`;
-    })
-    .join("\n");
+  const categoryRows = renderCategoryRows();
 
   return `# Verity — Security-Patched Container Images
